requirements.txt
opencv-contrib-python
numpy
pandas
dlib
imutils
scipy
pillow
